requirements.txt
pypdf==5.6.0
python-docx==1.1.2
spacy==3.8.7
